fix(desktop): drop the client-side follow-up queue
The renderer-local Zustand queue conflicted with SPEC 105: FR-027 requires
concurrent follow-ups to return a generic recoverable busy conflict and
states clients MUST NOT invent a local queue, FR-101 requires queueing to
be explicit rather than silent, and FR-100 requires pending messages to be
durable server-owned records.

The local queue also lost messages: the composer cleared the draft on
Enter, so a reload or window close before the active turn finished dropped
the follow-up with no trace, and queued items were invisible to the
mobile, browser, and CLI shells.

Every submit is now a direct send. A follow-up aimed at a busy thread
surfaces the existing safe thread_busy copy and keeps the draft. Restore a
runtime-owned durable queue before reintroducing any queueing UI.
